Implement input-output equilibrium calculation utilities functions
Implement the `calculate_output_equilibrium_setpoint` and
`calculate_input_equilibrium_setpoint` functions in
'initial_state_estimation.py'. These functions calculate equilibrium
input-output setpoints of a system, which is required for defining
the terminal conditions in the Data-Driven MPC controller and
initializing past input-output pairs for a given initial internal
state of a system.