Add terminal constraint argument to `DirectDataDrivenMPCController`
Add the `use_terminal_constraint` argument to define whether the
terminal equality constraint is used in the Data-Driven MPC formulation
for each controller created via `DirectDataDrivenMPCController`.

This addition allows creating controllers that omit this constraint,
enabling performance comparisons with the controller formulations
proposed by J. Berberich et al. in the referenced paper. It also
enables the reproduction of the example presented in the paper.