Remove output setpoint estimation from example script
- Remove system output setpoint override logic in example script,
as it does not reflect typical data-driven controller operation.
Previously, the controller's output setpoint `y_s` was overridden
with an estimate using system equations to avoid infeasible
solutions in Nominal Data-Driven MPC controllers. However, in a
real-world application, this would not be possible, since the
system model is assumed to be unknown.
- Update script docstring to reflect this change.