Enable animation frames reduction in plot animation
- Refactor `plot_input_output_animation` and `update_data_animation`
to display multiple data points per frame, reducing the total number
of required frames to display all the data. This setting is
defined by the new `points_per_frame` argument.
- Rename the `frame` argument to `index` in `update_data_animation`
to better describe that it represents the current data index.
- Update docstrings to reflect the changes.

This change allows the reduction of the number of animation frames
required to display input-output data, which enables the generation
of low FPS animation output files, such as GIFs.