Remove noise parsing argument from example script
- Remove the estimated system measurement noise (`eps_max`)
argument from the example script. Previously, it overrode the
`eps_max` parameter for both system (simulation) and controller,
which is not always correct, as this parameter should be
independent for each configuration. Now, `eps_max` can only be set
via the system and controller configuration files.
- Remove the `eps_bar` argument from
`get_data_driven_mpc_controller_params` since it is no longer used
to override the parameters of created controllers.

This change ensures that the `eps_max` parameter is handled
independently for both system (simulation) and controller through
their respective configuration files. This improves the clarity of
the configuration process in the example script.